moved tiled panes outside
